+++ askama_actix/src/lib.rs
@@ -6,11 +6,6 @@
 //! [the example](https://github.com/djc/askama/blob/master/testing/tests/actix_web.rs)
 //! from the Askama test suite for more on how to integrate.
 
-extern crate askama;
-extern crate actix_web;
-extern crate bytes;
-extern crate mime_guess;
-
 use std::fmt;
 
 struct BytesWriter {
@@ -43,7 +38,7 @@ impl fmt::Write for BytesWriter {
 // Older versions that don't have it exposed are easier this way. If ext is empty or no
 // associated type was found, then this returns `application/octet-stream`, in line with how
 // actix_web handles it in newer releases.
-pub use self::actix_web::{
+pub use actix_web::{
     error::ErrorInternalServerError, Error, HttpRequest, HttpResponse, Responder,
 };
 
@@ -61,5 +56,58 @@ impl<T: askama::Template> TemplateIntoResponse for T {
         Ok(HttpResponse::Ok()
             .content_type(ctype.as_str())
             .body(buffer.freeze()))
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use actix_web::http::header::CONTENT_TYPE;
+    use actix_web::test;
+    use actix_web::HttpMessage;
+    use askama::Template;
+    use bytes::Bytes;
+    use super::TemplateIntoResponse;
+
+    #[derive(Template)]
+    #[template(path = "../../testing/templates/hello.html")]
+    struct HelloTemplate<'a> {
+        name: &'a str,
+    }
+
+    #[test]
+    fn test_actix_web() {
+        let mut srv = test::TestServer::new(|app| app.handler(|_| HelloTemplate { name: "world" }));
+
+        let request = srv.get().finish().unwrap();
+        let response = srv.execute(request.send()).unwrap();
+        assert!(response.status().is_success());
+        assert_eq!(
+            response.headers().get(CONTENT_TYPE).unwrap(),
+            "text/html; charset=utf-8"
+        );
+
+        let bytes = srv.execute(response.body()).unwrap();
+        assert_eq!(bytes, Bytes::from_static("Hello, world!".as_ref()));
+    }
+
+    #[test]
+    fn test_actix_web_responder() {
+        let mut srv = test::TestServer::new(|app| {
+            app.handler(|_| {
+                let name = "world".to_owned();
+                HelloTemplate { name: &name }.into_response()
+            })
+        });
+
+        let request = srv.get().finish().unwrap();
+        let response = srv.execute(request.send()).unwrap();
+        assert!(response.status().is_success());
+        assert_eq!(
+            response.headers().get(CONTENT_TYPE).unwrap(),
+            "text/html; charset=utf-8"
+        );
+
+        let bytes = srv.execute(response.body()).unwrap();
+        assert_eq!(bytes, Bytes::from_static("Hello, world!".as_ref()));
     }
 }
